src/requirements.txt
PyYAML==5.4.1
cfn-tools==0.1.6
cfn-flip==1.2.3